Add files via upload
Job and Course Proposer is an application that’ll help the newly joined university fresher’s and fresh graduates. A list of courses are provided that’ll help students to increase their skill in that particular field and a list of jobs are provided with necessary requirements that both graduate and undergraduate students can avail if they meet the required requirements by forwarding their CV to them.